requirements.txt
# Core dependencies
torch>=1.13.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Graph processing
torch-geometric>=2.3.0
networkx>=2.6.0
h3>=3.7.0

# Optimization
gurobipy>=10.0.0  # Requires license

# Environment
gym>=0.21.0
pyyaml>=6.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional for visualization
plotly>=5.0.0
folium>=0.12.0

fastparquet
pyarrow
